# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Probity API",
  "description": "Evidence before payment.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/vendors": {
   "get": {
    "tags": [
     "vendors"
    ],
    "summary": "List Vendors",
    "operationId": "list_vendors_api_v1_vendors_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Q"
      }
     },
     {
      "name": "include_archived",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Archived"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Vendors Api V1 Vendors Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "vendors"
    ],
    "summary": "Create Vendor",
    "operationId": "create_vendor_api_v1_vendors_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VendorIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Vendor Api V1 Vendors Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/workspace/onboarding": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "Onboarding",
    "operationId": "onboarding_api_v1_workspace_onboarding_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Onboarding Api V1 Workspace Onboarding Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/workspace": {
   "patch": {
    "tags": [
     "workspace"
    ],
    "summary": "Rename Workspace",
    "operationId": "rename_workspace_api_v1_workspace_patch",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WorkspacePatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Rename Workspace Api V1 Workspace Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/imports/templates/{kind}": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "Import Template",
    "operationId": "import_template_api_v1_imports_templates__kind__get",
    "parameters": [
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "vendors",
        "invoices",
        "purchase_orders"
       ],
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/imports/spec": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "Import Spec",
    "operationId": "import_spec_api_v1_imports_spec_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Import Spec Api V1 Imports Spec Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/imports/{kind}": {
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Run Import",
    "description": "dry_run=true (default) validates and returns a report without writing. dry_run=false commits:\nrefused if any row is invalid unless skip_invalid=true (then only valid rows are written).",
    "operationId": "run_import_api_v1_imports__kind__post",
    "parameters": [
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "vendors",
        "invoices",
        "purchase_orders"
       ],
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "dry_run",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Dry Run"
      }
     },
     {
      "name": "skip_invalid",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Skip Invalid"
      }
     },
     {
      "name": "verification_note",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 1000
        },
        {
         "type": "null"
        }
       ],
       "title": "Verification Note"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_run_import_api_v1_imports__kind__post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Run Import Api V1 Imports  Kind  Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/imports": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "List Imports",
    "operationId": "list_imports_api_v1_imports_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Imports Api V1 Imports Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/cases/{case_id}/notes": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "List Notes",
    "operationId": "list_notes_api_v1_cases__case_id__notes_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Notes Api V1 Cases  Case Id  Notes Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Add Note",
    "operationId": "add_note_api_v1_cases__case_id__notes_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NoteIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Note Api V1 Cases  Case Id  Notes Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/notifications": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "List Notifications",
    "operationId": "list_notifications_api_v1_notifications_get",
    "parameters": [
     {
      "name": "unread_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Unread Only"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Notifications Api V1 Notifications Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/notifications/{notification_id}/read": {
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Mark Read",
    "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
    "parameters": [
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Notification Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Mark Read Api V1 Notifications  Notification Id  Read Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/notifications/read-all": {
   "post": {
    "tags": [
     "workspace"
    ],
    "summary": "Mark All Read",
    "operationId": "mark_all_read_api_v1_notifications_read_all_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Mark All Read Api V1 Notifications Read All Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/workspace/export": {
   "get": {
    "tags": [
     "workspace"
    ],
    "summary": "Export Workspace",
    "description": "Owner-only machine-readable export of the workspace's business data (bank numbers stay masked).",
    "operationId": "export_workspace_api_v1_workspace_export_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_run_import_api_v1_imports__kind__post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_run_import_api_v1_imports__kind__post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "NoteIn": {
    "properties": {
     "text": {
      "type": "string",
      "maxLength": 4000,
      "minLength": 1,
      "title": "Text"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "NoteIn"
   },
...
   "VendorIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 300,
      "minLength": 2,
      "title": "Name"
     },
     "gstin": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 15
       },
       {
        "type": "null"
       }
      ],
      "title": "Gstin"
     },
     "pan": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 10
       },
       {
        "type": "null"
       }
      ],
      "title": "Pan"
     },
     "address": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1000
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "website": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Website"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "VendorIn"
   },
...
   "WorkspacePatch": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 2,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "WorkspacePatch"
   }
  }
 }
}